re organise landing page
