+++ packages/datatrak-web/src/__tests__/mocks/browser.js
