+++ packages/datatrak-web/src/main.tsx
@@ -7,7 +7,7 @@ import { render as renderReactApp } from 'react-dom';
 import { App } from './App';
 
 if (process.env.NODE_ENV === 'development') {
-  const { worker } = await import('../mocks/browser');
+  const { worker } = await import('./__tests__/mocks/browser');
   worker.start();
 }
