+++ packages/datatrak-web/src/views/LandingPage/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { LandingPage } from './LandingPage';
